SchoolApp/SchoolApp/requirements.txt
Django
cryptography
django-encrypted-model-fields